morphemic_integration/forecasting_gluonts/morphemic/morphemic-performance-model/ml_code/src/requirements.txt
setuptools

numpy>=1.9.0
scipy>=0.14.1

joblib>=0.14.1
scikit-learn>=0.24.0,<0.25.0

dask
distributed>=2.2.0
pyyaml
pandas>=1.0
liac-arff

ConfigSpace>=0.4.14,<0.5
pynisher>=0.6.3
pyrfr>=0.7,<0.9
smac>=0.13.1,<0.14
pandas 
emcee 
pyDOE 
scikit-optimize 
sklearn 
auto-sklearn 
requests 
grpcio 
protobuf 
stomp.py
fastapi 
uvicorn